requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
boto3
